Move trash handling to separate module.
