+++ dttools/src/trash.h
@@ -1,0 +1,8 @@
+#ifndef TRASH_H
+#define TRASH_H
+
+void trash_setup( const char *trashdir );
+void trash_file( const char *filename );
+void trash_empty();
+
+#endif
